Added flyway and unit tests

+++ src/main/java/com/breeze/music/repository/MusicRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MusicRepository extends JpaRepository<Song, Integer> {
+public interface MusicRepository extends JpaRepository<Song, Long> {
     Optional<Song> getSongById(Long id);
 }
